feat: implement jwt provider and password hasher

+++ src/Authentication.Api/Responses/LogInResponse.cs
@@ -6,5 +6,5 @@ public class LogInResponse
 {
     public required string AccessToken { get; set; }
     public required string TokenType { get; set; }
-    public required int ExpiresIn { get; set; }
+    public required long ExpiresIn { get; set; }
 }
